Misc PR adjustments; restructure cron job + formatting.
