+++ src/index.ts
@@ -1,12 +1,11 @@
 import dotenv from "dotenv";
 
+import { scheduleCronJobs } from "./cron";
 import "./config/sentryConnection";
-import logger from "./config/loggingConfig";
 import { connectDB } from "./config/dbConnection";
 import app from "./utils/app";
 import { env } from "./utils/env";
-import {runSanctionBot} from "./services/sanctionBotJob.service";
-const schedule = require("node-schedule");
+import logger from "./config/loggingConfig";
 
 dotenv.config();
 
@@ -27,6 +26,8 @@ const startServer = async () => {
       });
     }
 
+
+
     logger.info("Server setup initiated.");
   } catch (error) {
     logger.error('Server failed to initialize:', error);
@@ -35,12 +36,7 @@ const startServer = async () => {
 
 // Start the server setup process
 startServer();
-
-// Schedule the job to run daily at 22:00 UTC using node-schedule.
-schedule.scheduleJob('0 22 * * * *', async () => {
-  logger.info('Scheduled job triggered at 22:00 UTC.');
-  await runSanctionBot().then(() => logger.info("Sanction Bot finished execution!"));
-  logger.info('Scheduled job finished running');
-});
+// Start the scheduled cron job(s)
+scheduleCronJobs();
 
 export default app;
